requirements.txt
supabase
streamlit
python-dotenv
MLB-StatsAPI